Extract svg part of LineChart
