+++ packages/palette/src/elements/LineChart/LineChart.tsx
@@ -1,18 +1,14 @@
-import { interpolateArray } from "d3-interpolate"
-import { line as d3Line } from "d3-shape"
 import React, { useEffect, useRef, useState } from "react"
-import { Spring } from "react-spring"
 import styled from "styled-components"
-import { color, space } from "../../helpers"
-import { max, useHasEnteredViewport } from "../../helpers/visualizationHelpers"
+import { space } from "../../helpers"
+import { useHasEnteredViewport } from "../../helpers/visualizationHelpers"
 import { coerceLabel } from "../BarChart"
 import { BarHoverLabel } from "../BarChart/Bar"
 import { BarLabelProps } from "../BarChart/BarLabel"
 import { ProvideMousePosition } from "../BarChart/MousePositionContext"
 import { Flex } from "../Flex"
 import { Sans } from "../Typography"
-import { Line } from "./Line"
-import { Point } from "./Point"
+import { LineChartSVG } from "./LineChartSVG";
 
 export interface PointDescriptor {
   value: number
@@ -33,8 +29,6 @@ const DEFAULT_HEIGHT = 87
  * Useful for visualizing a time series, etc.
  */
 export const LineChart = ({ points, chartHeight }: LineChartProps) => {
-  const maxValue: number = max(points, item => item.value)
-
   const [width, setWidth] = useState(0)
   const [height, setHeight] = useState(DEFAULT_HEIGHT)
   const [hoverIndex, setHoverIndex] = useState(-1)
@@ -65,21 +59,6 @@ export const LineChart = ({ points, chartHeight }: LineChartProps) => {
 
   const hasEnteredViewport = useHasEnteredViewport(wrapperRef)
 
-  const values = points.map(d => d.value)
-  const zeros = values.map(() => 0)
-  const valuesInterpolator = interpolateArray(zeros, values)
-
-  const w = width - 2 * margin
-  const h = height - 2 * margin
-
-  // maps value to x/y position
-  const displayYPosition = d => h - (d * h) / maxValue
-  const displayXPosition = (_d, i) => (i / (points.length - 1)) * w
-
-  const line = d3Line()
-    .x(displayXPosition)
-    .y(displayYPosition)
-
   return (
     <ProvideMousePosition>
       <Flex flexDirection="column" ref={wrapperRef as any} width="100%">
@@ -97,43 +76,14 @@ export const LineChart = ({ points, chartHeight }: LineChartProps) => {
                 </HoverHandler>
               ))}
             </Flex>
-            <Svg width={width} height={height}>
-              <g transform={`translate(0, ${margin})`}>
-                <line
-                  stroke={color("black10")}
-                  x1="0"
-                  x2={width}
-                  y1={h}
-                  y2={h}
-                />
-                <g transform={`translate(${margin}, 0)`}>
-                  <Spring
-                    from={{ num: 0 }}
-                    to={hasEnteredViewport ? { num: 1 } : { num: 0 }}
-                    delay={500}
-                  >
-                    {({ num }) => {
-                      const interpolatedValues: any = valuesInterpolator(num)
-                      return (
-                        <>
-                          <Line d={line(interpolatedValues as any)} />
-                          {interpolatedValues.map((value, index) => {
-                            return (
-                              <Point
-                                hovered={hoverIndex === index}
-                                key={index}
-                                cx={displayXPosition(value, index)}
-                                cy={displayYPosition(value)}
-                              />
-                            )
-                          })}
-                        </>
-                      )
-                    }}
-                  </Spring>
-                </g>
-              </g>
-            </Svg>
+            <LineChartSVG
+              width={width}
+              height={height}
+              margin={margin}
+              points={points}
+              hoverIndex={hoverIndex}
+              hasEnteredViewport={hasEnteredViewport}
+            />
             {points.filter(bar => bar.axisLabelX).length > 0 && (
               <Flex px="2" width={width}>
                 {points.map(({ axisLabelX }, i) => (
@@ -172,13 +122,9 @@ const HoverHandler = ({ children, index, hoverIndex, setHoverIndex }) => {
 const PointHoverArea = styled.div`
   flex: 1;
   z-index: 1;
-  margin-right: 1%;
+  margin-right: 1%; /* gap between area enabling tooptips */
   opacity: 0;
   transition: opacity 0.4s ease-in;
-`
-
-const Svg = styled.svg`
-  position: absolute;
 `
 
 const BarAxisLabelContainer = styled.div<AxisContainerProps>`
